% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubing.R
\name{cube}
\alias{cube}
\title{cubing a number input}
\usage{
cube(x)
}
\arguments{
\item{x}{A numeric value to be cubed}
}
\value{
cube of the input number
}
\description{
cubing a number input
}
